Siirrä adaptKuulutusYhteystiedot utiliin, pois asiakirjaAdapterista
