+++ backend/src/handler/asiakirjaAdapter.ts
@@ -1,22 +1,9 @@
-import {
-  AloitusKuulutusJulkaisu,
-  DBProjekti,
-  HyvaksymisPaatosVaiheJulkaisu,
-  KuulutusYhteystiedot,
-  NahtavillaoloVaiheJulkaisu,
-  Velho,
-  Yhteystieto,
-} from "../database/model";
+import { AloitusKuulutusJulkaisu, DBProjekti, HyvaksymisPaatosVaiheJulkaisu, NahtavillaoloVaiheJulkaisu, Velho } from "../database/model";
 import cloneDeep from "lodash/cloneDeep";
-import {
-  AloitusKuulutusTila,
-  HyvaksymisPaatosVaiheTila,
-  NahtavillaoloVaiheTila,
-  ProjektiRooli,
-} from "../../../common/graphql/apiModel";
+import { AloitusKuulutusTila, HyvaksymisPaatosVaiheTila, NahtavillaoloVaiheTila } from "../../../common/graphql/apiModel";
 import { deepClone } from "aws-cdk/lib/util";
-import { vaylaUserToYhteystieto } from "../util/vaylaUserToYhteystieto";
 import { findJulkaisuWithTila } from "../projekti/projektiUtil";
+import adaptKuulutusYhteystiedot from "../util/adaptStandardiYhteystiedot";
 
 function createNextAloitusKuulutusJulkaisuID(dbProjekti: DBProjekti) {
   if (!dbProjekti.aloitusKuulutusJulkaisut) {
@@ -102,34 +89,6 @@ export class AsiakirjaAdapter {
       return findJulkaisuWithTila(projekti.aloitusKuulutusJulkaisut, AloitusKuulutusTila.HYVAKSYTTY);
     }
   }
-}
-
-function adaptKuulutusYhteystiedot(
-  dbProjekti: DBProjekti,
-  kuulutusYhteystiedot: KuulutusYhteystiedot | null
-): Yhteystieto[] {
-  const yt: Yhteystieto[] = [];
-  const sahkopostit: string[] = [];
-  dbProjekti.kayttoOikeudet
-    .filter(
-      ({ kayttajatunnus, rooli }) =>
-        rooli === ProjektiRooli.PROJEKTIPAALLIKKO ||
-        kuulutusYhteystiedot?.yhteysHenkilot?.find((yh) => yh === kayttajatunnus)
-    )
-    .forEach((oikeus) => {
-      yt.push(vaylaUserToYhteystieto(oikeus));
-      sahkopostit.push(oikeus.email); //Kerää sähköpostit myöhempää duplikaattien tarkistusta varten.
-    });
-  if (kuulutusYhteystiedot.yhteysTiedot) {
-    kuulutusYhteystiedot.yhteysTiedot.forEach((yhteystieto) => {
-      if (!sahkopostit.find((email) => email === yhteystieto.sahkoposti)) {
-        //Varmista, ettei ole duplikaatteja
-        yt.push(yhteystieto);
-        sahkopostit.push(yhteystieto.sahkoposti);
-      }
-    });
-  }
-  return yt;
 }
 
 function adaptVelho(dbProjekti: DBProjekti): Velho {
